requirements.txt
einops==0.8.1
matplotlib==3.8.3
misc==0.0.0
mmengine==0.10.6
numpy==1.23.0
Pillow==12.0.0
thop==0.1.1.post2209072238
timm==1.0.22
torch==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.67.1
ultralytics_yolo==0.0.1
